+++ CHANGES.md
@@ -1,4 +1,5 @@
 #5.0.11 2016-??-??
+ - 2016-05-21 Fixed bug#[12012](http://bugs.otrs.org/show_bug.cgi?id=12012) - JSON Response contains all Dynamic Fields in Article response.
  - 2016-05-20 Fixed bug#[12069](http://bugs.otrs.org/show_bug.cgi?id=12069) - enabling/disabling Article Filter show wrong article.
  - 2016-05-19 7 Day Stats performance optimizations, thanks to Pawel Boguslawski.
  - 2016-05-20 Fixed bug#[12087](http://bugs.otrs.org/show_bug.cgi?id=12087) - AdminTemplate loses HTML tags.